feat(global states): move config to button base, small fixies
